scripts/requirements.txt
onnxruntime
pillow
numpy
